% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{focalExtract}
\alias{focalExtract}
\title{Focal extraction from a raster object
A function to extract raster values from neighbourhood of points.}
\usage{
focalExtract(
  r,
  p,
  neighbourhood = c(3, 3),
  fun = mean,
  na.rm = TRUE,
  progress = TRUE
)
}
\arguments{
\item{r}{raster layer}

\item{p}{the spatial point or sf objects}

\item{neighbourhood}{the dimention of the neighourhood kernel}

\item{fun}{function to summarize the extracted values (e.g. mean). If
no function is provided, a list is retuned}

\item{na.rm}{logical. If na.rm=TRUE (the default value), NA values are removed before fun is applied.}

\item{progress}{logical. Show a progress par.}
}
\value{

}
\description{
Focal extraction from a raster object
A function to extract raster values from neighbourhood of points.
}
