% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggResponse.R
\name{ggResponse}
\alias{ggResponse}
\title{Partial dependence plot by ggplot}
\usage{
ggResponse(
  models,
  covariates,
  colPlot = 3,
  responseName = "Prediction",
  index = 2,
  ...
)
}
\arguments{
\item{models}{a model object}

\item{covariates}{the covariates used in model fitting in raster or data.frame}

\item{colPlot}{integer. The number of colums for plotting}

\item{responseName}{character. the name for y axes}

\item{index}{integer. The columns used for prediction. This relates to the factor level for classification.}

\item{...}{other arguments e.g. type = 'response' in GLMs or type = 'prob' in randomForest or the number of trees in BRT/GBM.}
}
\value{

}
\description{
This function can be used to plot a partial dependence plot for any model that
can predict on data.frames.
}
\author{
Roozbeh Valavi
}
