% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggResponse.R
\name{ggResponse2}
\alias{ggResponse2}
\title{Title}
\usage{
ggResponse2(
  models,
  covariates,
  colPlot = 3,
  responseName = "Prediction",
  index = 2,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
