% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_helper.R
\name{glmnetPoly}
\alias{glmnetPoly}
\title{Orthogonal polynomial for glmnet}
\usage{
glmnetPoly(train, cols = NULL, all = TRUE)
}
\arguments{
\item{train}{training data.frame}

\item{cols}{Character. Column names to generate quadratic terms}

\item{all}{If TRUE, all the columns are used to generate quadratic terms}
}
\value{
A data.frame with polynomial terms of the original dataset.
}
\description{
A function to creat quadratic terms for lasso and ridge
}
