% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdata.R
\name{makePoints}
\alias{makePoints}
\title{Degitizing points on a spatial plot}
\usage{
makePoints(sp = TRUE, col = "red", ...)
}
\arguments{
\item{sp}{logical. To return spatial points}

\item{col}{the color to plot points with}

\item{...}{}
}
\value{

}
\description{
Degitizing points on a spatial plot
}
