% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myspatial.R
\docType{package}
\name{myspatial}
\alias{myspatial}
\title{myspatial: A collection of functions for spatial data analysis and modelling}
\description{
A collection of functions to facilitate processing spatial data for analysis and modelling.
The package is in its early stage of development, thus the functions provided here come with ABSOLUTELY NO WARRANTY.
}
\author{
Roozbeh Valavi
}
