% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_svm_raster}
\alias{predict_svm_raster}
\title{predict SVM to Raster}
\usage{
predict_svm_raster(r, model, factors = NULL, filename = NULL)
}
\arguments{
\item{r}{a raster file}

\item{model}{svm model from e1071 package}

\item{factors}{factor variables in the model}

\item{filename}{output file directory for the predicted map. If provided, the map
will be save on the disk.}
}
\value{
a raster file for prediction
}
\description{
This function is to predict the e1071 model object on raster data.
}
\author{
Roozbeh Valavi
}
