% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_svm_to_raster}
\alias{predict_svm_to_raster}
\title{predict SVM to Raster}
\usage{
predict_svm_to_raster(r, model, factors = NULL, filename = NULL)
}
\arguments{
\item{model}{svm model from e1071 package}
}
\value{

}
\description{
This function is to predict the e1071 model object on raster data.
}
