% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictSVMtoRaster}
\alias{predictSVMtoRaster}
\title{predictSVMtoRaster}
\usage{
predictSVMtoRaster(r, model, factors = NULL, filename = NULL)
}
\arguments{
\item{r}{raster file}

\item{model}{svm model from e1071 package}

\item{factors}{factor variables in the model}

\item{filename}{output file directory for the predicted map. If provided, the map
will be save in the disk.}
}
\description{
This function is to predict the e1071 model object on raster data.
}
