% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_helper.R
\name{quadratic_for_glmnet}
\alias{quadratic_for_glmnet}
\title{Orthogonal quadratic polynomials for glmnet}
\usage{
quadratic_for_glmnet(df, cols = NULL, all = TRUE)
}
\arguments{
\item{df}{The data.frame of training covariates}

\item{cols}{Character. Column names to generate quadratic terms}

\item{all}{If TRUE, all the columns are used to generate quadratic terms}
}
\value{
A data.frame with polynomial terms of the original dataset.
}
\description{
A function to creat quadratic terms for lasso and ridge
}
