% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rasterPlot}
\alias{rasterPlot}
\title{rasterPlot}
\usage{
rasterPlot(map, cols = terrain.colors(10), legend = TRUE)
}
\arguments{
\item{map}{raster of SDM probability}

\item{cols}{a string vector containing the colours code}

\item{legend}{logical. to show legend or not.}
}
\value{

}
\description{
Plot probability raster of species distribution models in ggplot,
with histogram of the values.
}
