% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppercatch.R
\name{upperCatchment}
\alias{upperCatchment}
\title{Extract and summarise raster values for upper-catchments}
\usage{
upperCatchment(ccm2, r, myfun = "mean")
}
\arguments{
\item{ccm2}{the CCM2 spatil polygon layer of subcatchments}

\item{r}{a Raster later or RasterStack for extracting values}

\item{myfun}{a function to summarise the values of input raster e.g. mean or sum}
}
\value{

}
\description{
Extract and summarise raster values for upper-catchments
}
