% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_glmnet_raster}
\alias{predict_glmnet_raster}
\title{predictglmnetRatser}
\usage{
predict_glmnet_raster(
  r,
  model,
  slambda = "lambda.min",
  quadratic = TRUE,
  trainingOriginalData = NULL,
  factors = NULL,
  filename = NULL
)
}
\arguments{
\item{r}{raster file}

\item{model}{glmnet model}

\item{slambda}{the lambda selection}

\item{quadratic}{logical. If the model was fitted with quadratic terms}

\item{trainingOriginalData}{original training data with only the predictor columns}

\item{factors}{factor variables in the model}

\item{filename}{output file directory for the predicted map. If provided, the map
will be save in the disk.}
}
\description{
This function is to predict the glmnet model object on raster data. It accepts
the quadratic transformation if it was provided in the original model fit.
}
