% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{rangeChange}
\alias{rangeChange}
\title{Range change}
\usage{
rangeChange(current, future, threshold = NULL)
}
\arguments{
\item{current}{the current raster map}

\item{future}{the future raster map}

\item{threshold}{optional. A threshold to change probability map to 0 and 1}
}
\value{

}
\description{
The values of raster show:
0 = unsiitable
1 = loss
2 = stable
3 = gain
}
